feat(helper): add aggregation logic to new helper
