+++ app/services/forest_liana/pie_stat_getter.rb
@@ -1,5 +1,6 @@
 module ForestLiana
   class PieStatGetter < StatGetter
+    include AggregationHelper
     attr_accessor :record
 
     def perform
@@ -42,44 +43,8 @@ module ForestLiana
       end
     end
 
-    def resolve_field_path(field_param, default_field = 'id')
-      return "#{@resource.table_name}.#{default_field}" unless field_param
-
-      if field_param.include?(':')
-        association, field = field_param.split ':'
-        associated_resource = @resource.reflect_on_association(association.to_sym)
-        "#{associated_resource.table_name}.#{field}"
-      else
-        "#{@resource.table_name}.#{field_param}"
-      end
-    end
-
     def groupByFieldName
       resolve_field_path(@params[:groupByFieldName])
-    end
-
-    def aggregation_sql(type, field)
-      field_path = resolve_field_path(field)
-
-      case type
-      when 'sum'
-        "SUM(#{field_path})"
-      when 'count'
-        "COUNT(DISTINCT #{field_path})"
-      else
-        raise "Unsupported aggregator : #{type}"
-      end
-    end
-
-    def aggregation_alias(type, field)
-      case type
-      when 'sum'
-        "sum_#{field.downcase}"
-      when 'count'
-        'count_id'
-      else
-        raise "Unsupported aggregator : #{type}"
-      end
     end
   end
 end
